@@ -1,12 +1,12 @@
 plugins {
-    id 'org.springframework.boot' version '2.4.7'
+    id 'org.springframework.boot' version '3.2.1'
     id 'io.spring.dependency-management' version '1.1.4'
     id 'java'
 }
 
 group = 'com.struchev.invest'
 version = '1.0.0'
-sourceCompatibility = '11'
+sourceCompatibility = '21'
 
 configurations {
     compileOnly {
@@ -34,9 +34,9 @@ dependencies {
     implementation 'org.springframework.retry:spring-retry:1.2.5.RELEASE'
     implementation 'org.springframework:spring-aspects:5.2.8.RELEASE'
 
-    implementation 'com.github.pengrad:java-telegram-bot-api:5.7.0'
+    implementation 'com.github.pengrad:java-telegram-bot-api:7.0.0'
 
-    implementation 'com.google.guava:guava:30.1.1-jre'
+    implementation 'com.google.guava:guava:33.0.0-jre'
 
     implementation 'org.reactivestreams:reactive-streams:1.0.3'
     implementation 'com.squareup.okhttp3:okhttp:4.12.0'
@@ -52,16 +52,16 @@ dependencies {
 
     implementation 'net.bull.javamelody:javamelody-spring-boot-starter:1.91.0'
 
-    compileOnly 'org.projectlombok:lombok'
-    annotationProcessor 'org.projectlombok:lombok'
+    compileOnly 'org.projectlombok:lombok:1.18.30'
+    annotationProcessor 'org.projectlombok:lombok:1.18.30'
 
-    runtimeOnly 'org.postgresql:postgresql:42.7.0'
+    runtimeOnly 'org.postgresql:postgresql'
 
-    testAnnotationProcessor 'org.projectlombok:lombok'
-    testCompileOnly 'org.projectlombok:lombok'
+    testAnnotationProcessor 'org.projectlombok:lombok:1.18.30'
+    testCompileOnly 'org.projectlombok:lombok:1.18.30'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.testcontainers:testcontainers:1.15.3'
+    testImplementation 'org.testcontainers:testcontainers:1.19.7'
     testImplementation 'org.testcontainers:postgresql:1.15.3'
 
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
